refactor(openomni): split background limit middleware
